add a ban list for words that don't fit what I'm looking for.  Need to investigate an issue where auctions are finishing with more items than they start with.  I think this is happening on Steep and Cheap's side of things

+++ src/co/mattg/snoop/Auction.java
@@ -6,8 +6,8 @@ public class Auction {
 	public int reduction;
 	public int quantityStart;
 	public int quantityFinish;
-	public int timestampStart;
-	public int timestampFinish;
+	public long timestampStart;
+	public long timestampFinish;
 
 	/*
 	 * Returns a new auction representing the given product
@@ -17,8 +17,8 @@ public class Auction {
 				   int reduction,
 				   int quantityStart,
 				   int quantityFinish,
-				   int timestampStart,
-				   int timestampFinish) {
+				   long timestampStart,
+				   long timestampFinish) {
 		
 		this.product = product;
 		this.price = price;
